requirements.txt
-i https://pypi.org/simple
asgiref==3.7.2 ; python_version >= '3.7'
django==4.2.2
django-bootstrap5==23.3
django-filter==23.2
gunicorn==20.1.0
python-decouple==3.8
setuptools==67.8.0 ; python_version >= '3.7'
sqlparse==0.4.4 ; python_version >= '3.5'
typing-extensions==4.6.3 ; python_version < '3.11'
